fix premature resolve when fallback is used
when the JSLoader loads and attempts to parse a file, but the Worker throws an error, prompting the JSLoader to use the fallback, the resolve() call is run anyways, marking the JSLoader as ready. The fallback load then overwrites the Worker, making further calls to JSLoader.prototype.execute() hang indefinitely.